Add recipe for rtree.
In addition to the regular package installation via setup.py this also
sets the variable DYLD_FALLBACK_LIBRARY_PATH (which only has an effect
on OS X). The reason this is needed is that if the libray
'libspatialindex' is installed via conda then it is not normally picked
up by the system unless DYLD_FALLBACK_LIBRARY_PATH is set correctly, and
as a result importing rtree will fail.